refactor gdbinit/lldbinit
